### 2024-10-15 #7354 @editorialbot
[PRE REVIEW]: SenSE: Community SAR ScattEring model
**Submitting author:** @McWhity (<a href="http://orcid.org/0000-0001-5278-8379">Thomas Weiß</a>)
**Repository:** https://github.com/McWhity/sense
**Branch with paper.md** (empty if default branch): master
**Version:** v0.2
