### 2024-10-16 01db37b
bug fix quoting error, include workflow for paper.pdf creation
