+++ .github/workflows/draft-pdf.yml
@@ -25,7 +25,7 @@ jobs:
           # This is the output path where Pandoc will write the compiled
           # PDF. Note, this should be the same directory as the input
           # paper.md
-          path: Joss_paper/paper.pdf
+          path: joss_paper/paper.pdf
       - name: Commit PDF to repository
         uses: EndBug/add-and-commit@v9
         with:
