### 2024-10-16 25e13c9
add check if readme.md did actually change
