+++ .github/workflows/draft-pdf.yml
@@ -31,4 +31,4 @@ jobs:
         with:
           message: '(auto) Paper PDF Draft'
           # This should be the path to the paper within your repo.
-          add: 'paper.pdf' # 'paper/*.pdf' to commit all PDFs in the paper directory
+          add: 'joss_paper/paper.pdf' # 'paper/*.pdf' to commit all PDFs in the paper directory
