+++ joss_paper/paper.bib
@@ -184,6 +184,7 @@
 	address = {Ann Arbor},
 	title = {Microwave radar and radiometric remote sensing},
 	isbn = {978-0-472-11935-6},
+	doi = {10.3998/0472119356},
 	language = {eng},
 	publisher = {University of Michigan press},
 	author = {Ulaby, Fawwaz Tayssir and Long, David G.},
