### 2024-10-24 #7354 @McWhity
@crvernon you are right the original repository is [pygeo/sense](https://github.com/pygeo/sense). It was started by my former supervisor and colleague, Alexander Löw, who unfortunately died in a car accident. Thus, we do not have writing access to pygeo/sense, which is unfortunate. To continue the repository in his honor, another colleague forked it to [PMarzahn/sense](https://github.com/PMarzahn/sense). However, since he did not actually work on the Python code, I moved the repository to my account to continue updating and improving it. I know it is not ideal, but we had to deal with the circumstances. Alexander Löw (pygeo) is credited as author, and he is also acknowledged as the person who initiated the repository (see Paper draft).
