### 2024-11-18 #7354 @mbobra
👋 @McWhity Thanks for your submission! If you have any suggestions for reviewers, please let me know here (you can give me the Github handle without @ mentioning them, e.g. mbobra instead of @mbobra). Feel free to look through [our database](https://reviewers.joss.theoj.org/reviewers) of reviewers. You can also suggest people outside of this database.
