### 2024-12-21 #7612 @editorialbot
[REVIEW]: SenSE: Community SAR ScattEring model
**Submitting author:** @McWhity (<a href="http://orcid.org/0000-0001-5278-8379">Thomas Weiß</a>)
**Repository:** https://github.com/McWhity/sense
**Branch with paper.md** (empty if default branch): master
**Version:** v1.0.0
