### 2024-12-27 #7612 @arthur-e
@McWhity, re: "Contribution and authorship," I am noting that this repository is a fork of another repository, which is unusual for a JOSS submission. Can you clarify:

1. What is the status of the upstream repository [PMarzahn/sense](https://github.com/PMarzahn/sense)? Could PMarzahn update the README to point to the new repository? (Would that be appropriate?)
2. Who is PMarzahn and should they be added to the list of authors? They have contributed a low number of commits but they are the owner of the upstream repository.
3. Should [bulli92](https://github.com/bulli92) be added to the authors list? They have contributed a large number of commits (101 to your 76 in the downstream repository).
