### 2024-12-27 #7612 @scottstanie
Hi @arthur-e , I think I saw the explanation for the forked repository in the Pre Review: https://github.com/openjournals/joss-reviews/issues/7354#issuecomment-2434577712
