### 2024-12-28 #7612 @arthur-e
@McWhity I have verified the Jupyter Notebooks run as expected; thanks for providing those. However, I feel they can be substantially improved to help end-users. The [`sm_retrieval`](https://github.com/McWhity/sense/blob/master/docs/notebooks/sm_retrieval.ipynb) is particularly compelling but could use: 1) A better narrative throughout, i.e., more details as to each step in the process of simulating backscatter and retrieving soil moisture; 2) More code comments; 3) Brief, high-level descriptions of some alternative parameters or alternative inputs, to illustrate the range of options and to link to the documentation.

As a potential end-user but not a specialist, I'd love to bring, e.g., a Sentinel-1 backscatter image to this framework and retrieve soil moisture or vegetation optical depth. Can you include an additional notebook that shows how to work with real data? The purposes of the remaining three notebooks (i.e., other than `sm_retrieval`) are unclear and seem unconnected to potential applications. They, too, could use more detail and better narrative.

Similarly, I think the JOSS manuscript could use a brief statement (circa one paragraph) on the available SAR imaging platforms (air-borne or space-borne); not an exhaustive list, but perhaps mentioning some of the more prevalent options and how those data could (or could not) be used with this framework. What is required to use data from those platforms with this framework?
