### 2024-12-29 #7612 @McWhity
First of all. Thank you @arthur-e and @scottstanie for your time and effort to help me improve the repository during the JOSS review process.

Regarding the authorship:

>    112	Thomas Weiß
>     98	Alex
>     72	McWhity
>      7	Ben
>      6	PMarzahn
>      3	loewalex
>      2	BenMGeo
>      1	P.Marzahn

- In retrospective it would have been better if I had directly forked the original repository [pygeo/sense](https://github.com/pygeo/sense). However, as that was not the case and [PMarzahn/Sense](https://github.com/PMarzahn/sense) will no longer be updated I will ask [PMarzahn](https://github.com/PMarzahn) to update the README to point to the repository currently under review.
-  Alex, loewalex and [bulli92](https://github.com/bulli92) are actually the same person (Alexander Löw). He is listed as an author and he is again mentioned in the acknowledgements of the paper.
- Thomas Weiß and McWhity are the same person as well. I apologize for any confusion caused by the inconsistent use of names; back then, I was just beginning to use Python and Git.
- Additional commits were made by Ben (also associated with BenMGeo) and PMarzahn (also associated with P.Marzahn). These contributions were primarily focused on testing the repository's functionality shortly after the passing of [bulli92](https://github.com/bulli92).

Regarding the notebooks:

- @arthur-e: Thanks for testing
- I will incoperate your remarks in the new year
