### 2025-08-13 #7612 @McWhity
Since not everything has been fully revised yet, I suggest waiting at least until the end of the week for the remaining updates.

#### Comments @scottstanie

> #### Walkthrough 1
> For the plots in the initial usage, it looks like we're setting radar parameters, then plugging in a linspace of moisture. But then the output Sigma_VV is on the x-axis, and mv is on the y-axis. Is it supposed to be an inversion? Or is the plot flipped?

- You are right. The plot is just flipped. I changed it.

>
> #### Walkthrough 2
> for t in theta:
>     print(t)
>     I1 = I2EM(f, eps, s, l, t, acf_type='gauss', xpol=xpol, auto=auto)
>     I2 = I2EM(f, eps, s, l, t, acf_type='exp15', xpol=xpol, auto=auto)
> Is it not currently possible to pass the array of theta? The paper mentions that one of the advantages is in "processing large datasets": is there any example of what that might look like? I would have thought you'd need to do array operations for that.
> Also, is there some example of why/how you would use a "large scale" version of the code here? I've only breifly looked at the Ulaby, 2014 book, so I dont know how these models would be used at a larger scale.
> Either an example, or even a description in the paper would be helpful to me.

- Will be addressed during updating the notebooks

> #### Citations
> In the sphinx docs, the references appear like Oh92 [@oh_empirical_1992] - is it supposed to turn into a link?
> The citation syntax [@oh_empirical_1992] isn't rendering as a link on the intro page, while it works correctly on the Theory page. I can't actually see why the 01_intro page isn't working, it looks you used like the same style.

- That seems to be a problem of RtD handling numbered style bibliography on different pages. I added a second .bib file (now each page has its own bibliography) and used the option :cited: (only actual cited entries are shown). Problem should be solved.

> #### Dependencies/environment
> The environment currently is
> 	- scipy
> 	- numba
> 	- numpy
> 	- matplotlib
> 		I see there are a lot of plotting helper scripts
> 		Perhaps you might consider [matlotlib-base](https://stackoverflow.com/questions/61325817/differences-between-matplotlib-and-matplotlib-base) as a conda dependency to avoid making 		Qt a requirement to run the scattering models?
> For the following:
> 	- sphinxcontrib-bibtex>=0.4 # docs
> 	- pytest>=4.6 # testing
> 	- jupyter
> These are for testing/documentation building, so perhaps you could remove from the main requirement file (and I see you've already got the docs/requirements.txt). Not critical, just seems nice for users who stictly want to run the models in their code but may not want to install all the extra Jupyter dependencies.
> For
> 	- anyio*
> 	- future

- updated but I need to test if everything is working correctly

> #### General maintenance code things
> Thoughts on using black or ruff to auto-lint the code? Some of the model lines are stretched very long (200+ characters). If you're concerned about it being a large commit, you can add a git-blame-ignore-revs [file](https://gist.github.com/kateinoigakukun/b0bc920e587851bfffa98b9e279175f2).

- Changed

> #### On the Models
> I see the base model for a class looks like
> class SurfaceScatter(object):
>     def __init__(self, eps=None, ks=None, theta=None, kl=None, mv=None, C_hh=None, C_vv=None, C_hv=None, D_hh=None, D_vv=None, D_hv=None, **kwargs):
>         self.eps = eps
>         self.ks = ks
>         self.theta = theta
>         self.kl = kl
>         self.mv = mv
>         self.C_hh = C_hh
>         self.C_vv = C_vv
>         self.D_hh = D_hh
>         self.D_vv = D_vv
>         self.C_hv = C_hv
>         self.D_hv = D_hv
>         self._check()
>
>     def _check(self):
>         pass
>         # assert isinstance(self.eps, complex)
> Type hints would be helpful to clarify which parameters are scalars vs vectors/matrices, especially for parameters like eps and theta that might accept either.
> Also noticing that all the classes have (object), which hasn't been needed since ~python3.3. Is that on purpose to keep that to support something older? Or just hasn't been updated?
> Another way to make it easier to documentation easier for, e.g.
> class Scatterer(object):
>     def __init__(self, **kwargs):
>         # NOTE THAT THE arguments are not necessarily the
>         # particle scattering cross sections!!!!
>         # need to be made in a clearer way !!!!
>         self.sigma_s_hh = kwargs.get('sigma_s_hh', None)  # particle scattering cross area
>         assert self.sigma_s_hh is not None, 'Particle HH scattering cross section needs to be specified [m**-2]'
>         self.sigma_s_vv = kwargs.get('sigma_s_vv', None)  # particle scattering cross area
>         assert self.sigma_s_vv is not None, 'Particle VV scattering cross section needs to be specified [m**-2]'
>         self.sigma_s_hv = kwargs.get('sigma_s_hv', None)  # particle scattering cross area
>         assert self.sigma_s_hv is not None, 'Particle HV scattering cross section needs to be specified [m**-2]'
> is to use data classes, which throw and error if you dont pass the required arguments
> @dataclass
> class Scatterer:
>     """Particle HH scattering cross section"""
>         sigma_s_hh: float
>         """...(more documentation)"""
>         sigma_s_vv: float
>         """...(more documentation)"""
>         sigma_s_vh: float

- need to be done
