### 2025-08-15 #7612 @McWhity
### Done this week

- test new environment/dependency guidelines
- update code of entire python code – implement dataclass type
- update of jupyter notebooks started

### Next steps (continue next week)
- continue update jupyter notebooks (include comments of arthur-e)
- include available SAR imaging information in JOSS manuscript
- update README of Marzahn/sense
