### 2025-09-06 #7612 @McWhity
### Already done

  • Plot problem solved (switch of axes)
  • Citation problem solved
  • Update environment/dependency
  • Update code of entire python code – implement dataclass type
  • Clean code line length using ruff
  • Update of README for Marzahn/sense (link to new repository)
  • Include available SAR imaging information in JOSS manuscript
  • Update of jupyter notebooks started

### Only missing item

  • Finalize jupyter notebooks

Next week I will be on vacation. I will continue working on it from 16th of September onward.
