### 2025-12-28 #7612 @McWhity
The only missing part of the revision are the notebooks. I plan to revise the notebooks by 16th of January 2026.
