### 2026-01-19 #7612 @McWhity
@crvernon I was able to revise 1 notebook out of 3. Currently I am working on the 2nd notebook which should be finalized by the end of the week.
