### 2026-01-25 #7612 @McWhity
@crvernon , @mbobra : I really appreciate the time that was given to me and I apologize for the delays. I finished now the revision of the three original jupyter notebooks. Thus, the revision is almost complete. Until the end of next week (01.02.2026) I want to write a (new) last jupyter notebook (request from @arthur-e) that shows how to work with a real data set. Furthermore, I may need to check if there are some new errors based on the substantial change in the code. If I am not able to finish this by (01.02.2026) I/you can withdraw the submission. Thank you for your patience.
