### 2026-02-01 #7612 @McWhity
@crvernon , @mbobra , @scottstanie , @arthur-e
I finally revised the package based on your comments. The revisions are all merged to the master branch.

### **Comments @scottstanie **
---------------------------------------------------

>
> Walkthrough 1
> For the plots in the initial usage, it looks like we're setting radar parameters, then plugging in a linspace of moisture. But then the output Sigma_VV is on the x-axis, and mv is on the y-axis. Is it supposed to be an inversion? Or is the plot flipped?

- You are right. The plot is just flipped. I changed it. However, I changed a huge part of the notebook based on the comments from @arthur-e

> Walkthrough 2
> for t in theta:
>     print(t)
>     I1 = I2EM(f, eps, s, l, t, acf_type='gauss', xpol=xpol, auto=auto)
>     I2 = I2EM(f, eps, s, l, t, acf_type='exp15', xpol=xpol, auto=auto)
> Is it not currently possible to pass the array of theta? The paper mentions that one of the advantages is in "processing large datasets": is there any example of what that might look like? I would have thought you'd need to do array operations for that.
> Also, is there some example of why/how you would use a "large scale" version of the code here? I've only breifly looked at the Ulaby, 2014 book, so I dont know how these models would be used at a larger scale.
> Either an example, or even a description in the paper would be helpful to me.

- Unfortunately, it is not possible to pass an array of theta right now. This is currently not done yet. With the restructuring of the other models (your other comment regarding @dataclass) the I2EM model is not working right now. I plan to update this in the next weeks.

>
> Citations
> In the sphinx docs, the references appear like Oh92 [@oh_empirical_1992] - is it supposed to turn into a link?
> The citation syntax [@oh_empirical_1992] isn't rendering as a link on the intro page, while it works correctly on the Theory page. I can't actually see why the 01_intro page isn't working, it looks you used like the same style.

- That seems to be a problem of RtD handling numbered style bibliography on different pages. I added a second .bib file (now each page has its own bibliography) and used the option :cited: (only actual cited entries are shown). Problem should be solved.

> Dependencies/environment
> The environment currently is
> 	- scipy
> 	- numba
> 	- numpy
> 	- matplotlib
> 		I see there are a lot of plotting helper scripts
> 		Perhaps you might consider [matlotlib-base](https://stackoverflow.com/questions/61325817/differences-between-matplotlib-and-matplotlib-base) as a conda dependency to avoid making 		Qt a requirement to run the scattering models?
> For the following:
> 	- sphinxcontrib-bibtex>=0.4 # docs
> 	- pytest>=4.6 # testing
> 	- jupyter
> These are for testing/documentation building, so perhaps you could remove from the main requirement file (and I see you've already got the docs/requirements.txt). Not critical, just seems nice for users who stictly want to run the models in their code but may not want to install all the extra Jupyter dependencies.
> For
> 	- anyio*
> 	- future

- changed matplotlib to matplotlib-base for conda
- Put RtD/testing specific requirements in a separate “docs_requirements.txt” file.
- Update installation guide to point out that additional packages need to be installed for testing, RtD creation and using jupyter notebooks. However, for the docker still all requirements (including testing, RtD, jupyter) will be installed.
- Thank you for noticing. anyio and future are not needed anymore

>
> General maintenance code things
> Thoughts on using black or ruff to auto-lint the code? Some of the model lines are stretched very long (200+ characters). If you're concerned about it being a large commit, you can add a git-blame-ignore-revs [file](https://gist.github.com/kateinoigakukun/b0bc920e587851bfffa98b9e279175f2).

- changed

> On the Models
> I see the base model for a class looks like
>    class SurfaceScatter(object):
>       def __init__(self, eps=None, ks=None, theta=None, kl=None, mv=None, C_hh=None, C_vv=None, C_hv=None, D_hh=None, D_vv=None, D_hv=None, **kwargs):
>           self.eps = eps
>           self.ks = ks
>           self.theta = theta
>           self.kl = kl
>           self.mv = mv
>           self.C_hh = C_hh
>           self.C_vv = C_vv
>           self.D_hh = D_hh
>           self.D_vv = D_vv
>           self.C_hv = C_hv
>           self.D_hv = D_hv
>           self._check()
>
>     def _check(self):
>         pass
>         # assert isinstance(self.eps, complex)
> Type hints would be helpful to clarify which parameters are scalars vs vectors/matrices, especially for parameters like eps and theta that might accept either.
> Also noticing that all the classes have (object), which hasn't been needed since ~python3.3. Is that on purpose to keep that to support something older? Or just hasn't been updated?
> Another way to make it easier to documentation easier for, e.g.
> class Scatterer(object):
>     def __init__(self, **kwargs):
>         # NOTE THAT THE arguments are not necessarily the
>         # particle scattering cross sections!!!!
>         # need to be made in a clearer way !!!!
>         self.sigma_s_hh = kwargs.get('sigma_s_hh', None)  # particle scattering cross area
>         assert self.sigma_s_hh is not None, 'Particle HH scattering cross section needs to be specified [m**-2]'
>         self.sigma_s_vv = kwargs.get('sigma_s_vv', None)  # particle scattering cross area
>         assert self.sigma_s_vv is not None, 'Particle VV scattering cross section needs to be specified [m**-2]'
>         self.sigma_s_hv = kwargs.get('sigma_s_hv', None)  # particle scattering cross area
>         assert self.sigma_s_hv is not None, 'Particle HV scattering cross section needs to be specified [m**-2]'
> is to use data classes, which throw and error if you dont pass the required arguments
> @dataclass
> class Scatterer:
>     """Particle HH scattering cross section"""
>         sigma_s_hh: float
>         """...(more documentation)"""
>         sigma_s_vv: float
>         """...(more documentation)"""
>         sigma_s_vh: float

- Thank you for pointing this out. Yes the classes with (object) were not updated. The reviewed the entire python package and updated the code introducing the suggested @dataclass where applicable. More information on which type the parameter have are now provided as well. The only model that was not changed yet is the I2EM. Unfortunately this model need some more changes. The problem that an array of theta cannot be ingested need to be fixed as well.

### **Comments @arthur-e **
---------------------------------------------------

>  I have verified the Jupyter Notebooks run as expected; thanks for providing those. However, I feel they can be substantially improved to help end-users. The [sm_retrieval](https://github.com/McWhity/sense/blob/master/docs/notebooks/sm_retrieval.ipynb) is particularly compelling but could use: 1) A better narrative throughout, i.e., more details as to each step in the process of simulating backscatter and retrieving soil moisture; 2) More code comments; 3) Brief, high-level descriptions of some alternative parameters or alternative inputs, to illustrate the range of options and to link to the documentation.

 - Jupyter notebooks have been restructured and more comments were added.
 - Example 1 shows how to used surface model (forward modeling, output backscatter)
 - Example 2 shows how to combine surface and canopy model (forward modeling, output backscatter)
 - Example 3 shows how to retrieve soil moisture with a synthetic dataset (inversion, input backscatter)

> As a potential end-user but not a specialist, I'd love to bring, e.g., a Sentinel-1 backscatter image to this framework and retrieve soil moisture or vegetation optical depth. Can you include an additional notebook that shows how to work with real data? The purposes of the remaining three notebooks (i.e., other than sm_retrieval) are unclear and seem unconnected to potential applications. They, too, could use more detail and better narrative.

- I added a fourth jupyter notebook with the purpose to show how to use the package to retrieve soil moisture based on a real dataset (Sentinel-1 and Sentinel-2).

> Similarly, I think the JOSS manuscript could use a brief statement (circa one paragraph) on the available SAR imaging platforms (air-borne or space-borne); not an exhaustive list, but perhaps mentioning some of the more prevalent options and how those data could (or could not) be used with this framework. What is required to use data from those platforms with this framework?

- Paragraph was added to the end of Statement of Need section

---------------------------------------------------

@scottstanie @arthur-e - I really appreciated your reviews. Thank you. I will further try to improve this packages.
