### 2026-02-13 #7612 @arthur-e
Thanks for these updates @McWhity, they sound like substantial improvements. I can see that the repository has been updated but I would like to see:

- Could you please update the ReadTheDocs documentation as well, so it points to the new Examples you created?
- Could you please add the new paragraph (that was just added to the repo's Statement of Need) to the JOSS article (and re-generate)? The JOSS manuscript (PDF) hasn't been updated since 2024.

Finally, there was a question about authorship that was overlooked in the intervening time. I did see the explanation about `PMarzahn`, However, [from the Contributors,](https://github.com/McWhity/sense/graphs/contributors) it looks like there is still one author who isn't accounted for, `bulli92`. Is this a duplicate account for someone already listed as an author? Thanks!
