### 2026-03-18 #7612 @McWhity
@arthur-e Thank you for your feedback.

Regarding:
---

> Could you please update the ReadTheDocs documentation as well, so it points to the new Examples you created?

You are right. The ReadTheDocs documentation was only updated for the dev branch. I updated the "latest" and the "master" branch as well. Thank you for noticing.

> Could you please add the new paragraph (that was just added to the repo's Statement of Need) to the JOSS article (and re-generate)? The JOSS manuscript (PDF) hasn't been updated since 2024.

Again thank you for noticing. I included the new paragraph in the ReadTheDocs part. The updated version should now be available on ReadTheDocs as well as in the JOSS mauscript. The editorialbot did create the current version of the article as well.

> Finally, there was a question about authorship that was overlooked in the intervening time. I did see the explanation about PMarzahn, However, [from the Contributors,](https://github.com/McWhity/sense/graphs/contributors) it looks like there is still one author who isn't accounted for, bulli92. Is this a duplicate account for someone already listed as an author?

I know the author list is a little bit confusing. This is unfortunate as the person who original started this repository died during a car accident. Here is author list of the software report.

>    112	Thomas Weiß
    98	Alex
    72	McWhity
     7	Ben
     6	PMarzahn
     3	loewalex
     2	BenMGeo
     1	P.Marzahn

I am Thomas Weiß (name) and McWhity (GitHub username)
The original founder of the repository is Alex (name) and loewalex (GitHub username) and bulli92 (GitHub username). As he suddenly died in an car accident nobody has access to his accounts.
Ben (name) and BenMGeo (GitHub username) is the same person
PMarzahn and P.Marzahn is also the same person

I know the situation is not ideal but I hope this clarifies your question.

@scottstanie: Did you have time to look at the changes.
