### 2026-03-22 #7612 @scottstanie
Hi @McWhity , sorry for missing the earlier notifications on this.

Yes I think that all the new updates to the repo, and especially the real world notebook example, make the package much nicer to browse through and understand. thanks for adding those! this will be especially interesting to try out on new NISAR data and compare to the premade products.

I've updated my checklist above and think we can approve the paper.
