### 2026-03-27 #7612 @mbobra
Congratulations, @McWhity! We're almost ready to publish the paper. Could you please cut a new release of `SenSE`, [deposit it in Zenodo](https://guides.github.com/activities/citable-code/), and put the link to the Zenodo deposit in this thread? Please make sure that the author list and title on the Zenodo deposit matches the author list and title on the JOSS paper. Then I can move forward with accepting the paper.
