### 2026-03-27 #7612 @mbobra
Thank you both so very much for your thorough reviews, @scottstanie and @arthur-e! I personally cannot thank you enough for keeping tabs on this for a year and a half. If there's any way JOSS can help you (whether it is connecting further to the open source community or to open source organizations), please reach out. Your volunteer time keeps JOSS running and the JOSS team appreciates it very much ☀️
