### 2026-04-01 #7612 @McWhity
@mbobra: I did cut a new release v1.0.0. The release (v1.0.0) is deposited in Zenodo. The Zenodo link is [https://zenodo.org/records/19376069](https://zenodo.org/records/19376069). I checked the release number, title and author list/details on Zenodo. They should match the JOSS paper. Thank you for your help and support.

Let me know if there are further things that need to be done.
