### 2026-04-07 #7612 @McWhity
> Thank you for your submission and all the edits, [@McWhity](https://github.com/McWhity)! It was a long road, but we got there in the end. I hope you feel it improved the code. This is a really nice paper!

@mbobra : It certainly was a long road! I’m very grateful for the time I got to improve the code. The package is much better now thanks to the JOSS review process. I highly appreciate the time and effort the editors and reviewers have invested. Thank you.
