### 2026-04-22 #7612 @crvernon
@McWhity - This submission is looking great.  I just need you to make the following changes to the paper before I accept for publication:

- [x] LINE 21-22:  The in-text citation here should be formatted as "Dobson et al. (1985)" instead of "Dobson et al. (Dobson et al., 1985)". You can do this in your paper.md file by changing "...model by Dobson et al. [@dobson_microwave_1985]" to "...model by @dobson_microwave_1985"
- [x] LINE 61:  Similar in-text citation formatting problem here as well.  "Ulaby and Long (Ulaby & Long, 2014)" should be formatted as "Ulaby and Long (2014)"

No new release or archive is necessary after making these changes.

Let me know when this is done and I'll accept.  Thank you.
