### 2026-04-23 #7612 @McWhity
@crvernon I updated:

> - [x] LINE 21-22:  The in-text citation here should be formatted as "Dobson et al. (1985)" instead of "Dobson et al. (Dobson et al., 1985)". You can do this in your paper.md file by changing "...model by Dobson et al. [@dobson_microwave_1985]" to "...model by @dobson_microwave_1985"
> - [x] LINE 61:  Similar in-text citation formatting problem here as well.  "Ulaby and Long (Ulaby & Long, 2014)" should be formatted as "Ulaby and Long (2014)"

as requested.

However, using @ulaby_microwave_2014 results in "Ulaby **&** Long (2014)" and not in "Ulaby **and** Long (2014). Is this a problem?

Furthermore, I found another citation error which I fixed. Line 19/20 is now showing the correct citations "(Roo et al., 2001; Ulaby & Long, 2014)". I had to change "**,**" to "**;**" within the paper.md file.

Let me know if anything else need to be changed. Thank you.
